(#23116) Make SSL Validator API more explicit
This refactors the API with an abstract superclass that contains
the API documentation and two factory methods; to obtain the no
validation implementation, and the default implementation.

Yardoc reworked.
