+++ lib/puppet/ssl.rb
@@ -7,5 +7,6 @@ module Puppet::SSL # :nodoc:
   require 'puppet/ssl/host'
   require 'puppet/ssl/oids'
   require 'puppet/ssl/validator'
-  require 'puppet/ssl/no_validator'
+  require 'puppet/ssl/validator/no_validator'
+  require 'puppet/ssl/validator/default_validator'
 end
