+++ lib/puppet/ssl/validator/no_validator.rb
@@ -1,6 +1,8 @@
-# Perform no SSL verification
+# Performs no SSL verification
 # @api private
-class Puppet::SSL::NoValidator
+#
+class Puppet::SSL::Validator::NoValidator < Puppet::SSL::Validator
+
   def setup_connection(connection)
     connection.verify_mode = OpenSSL::SSL::VERIFY_NONE
   end
